+++ python-libraries/data-platform-catalogue/data_platform_catalogue/__init__.py
@@ -1,5 +1,5 @@
-from .client import CatalogueClient  # noqa: F401
+from .client import DataHubCatalogueClient  # noqa: F401
+from .client import OpenMetadataCatalogueClient  # noqa: F401
 from .client import CatalogueError, ReferencedEntityMissing  # noqa: F401
-from .entities import CatalogueMetadata  # noqa: F401
 from .entities import DataProductMetadata  # noqa: F401
-from .entities import TableMetadata  # noqa: F401
+from .entities import CatalogueMetadata, DataLocation, TableMetadata  # noqa: F401