+++ terraform/github/analytical-platform-repositories.tf
@@ -1,5 +1,16 @@
 locals {
   analytical_platform_repositories = {
+    "analytical-platform" = {
+      name         = "analytical-platform"
+      description  = "Analytical Platform"
+      topics       = ["ministryofjustice", "analytical-platform"]
+      has_issues   = false
+      homepage_url = "https://user-guidance.analytical-platform.service.justice.gov.uk"
+      access = {
+        admins  = [module.data_platform_teams["data-platform-apps-and-tools"].id]
+        pushers = [module.data_platform_team.id]
+      }
+    }
     "analytics-platform-infrastructure" = {
       name                                   = "analytics-platform-infrastructure"
       description                            = "Analytical Platform Infrastructure"